wrap call to stitch each individual recipe in a try...except block
gridded stitching Fails randomly on pic - it's not the code, or data. If I download the gridded recipes, I can stitch the gridded data locally no problem, even the one it fails on.
Wrapping in a try...except so that it just skips those recipes and prints them to the slurm.out file, we can go back and stitch the handful of netcdfs if needed after. Better than babysitting pic runs.